refactor flow init node, add utils for digital signature, sha256

+++ tendermint/consensus/select_proposer.go
@@ -1,12 +1,12 @@
 package consensus
 
 type ProposerSelector struct {
-	NodeIds []int
+	NodeIds []int32
 	count int
 }
 
-func (selector *ProposerSelector) SelectProposer(height, round int32) int{
-	idx := int32(selector.count % len(selector.NodeIds))
+func (selector *ProposerSelector) SelectProposer(height, round int32) int32{
+	idx := selector.count % len(selector.NodeIds)
 	selector.count++
 	return selector.NodeIds[idx]
 }
